==> notebooks/0-Base.ipynb <==
%load_ext autoreload

%autoreload 2

import numpy as np

import pandas as pd

import matplotlib.pyplot as plt

%matplotlib inline

%config InlineBackend.figure_format = 'retina'

import seaborn as sns

from tqdm.auto import tqdm

DATA_PATH = '../data'
