==> notebooks/1-Data-splitting.ipynb <==
%run 0-Base.ipynb

from elo_competition.utils import split_by_columns

split_by_columns(f'{DATA_PATH}/raw/historical_transactions.csv',
                 f'{DATA_PATH}/1-splitted/historical_transactions')

split_by_columns(f'{DATA_PATH}/raw/new_merchant_transactions.csv',
                 f'{DATA_PATH}/1-splitted/new_merchant_transactions')
